refactor: currency repository find all method removed and find by code method turned to find by id in

+++ src/main/java/com/utkucan/fxapp/application/service/ExchangeRateSyncService.java
@@ -4,10 +4,10 @@ import com.utkucan.fxapp.domain.enums.CurrencyCode;
 import com.utkucan.fxapp.domain.entity.Currency;
 import com.utkucan.fxapp.domain.repository.CurrencyRateProvider;
 import com.utkucan.fxapp.domain.repository.CurrencyRepository;
-import jakarta.annotation.PostConstruct;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -35,7 +35,7 @@ public class ExchangeRateSyncService {
             Set<Currency> currencySet = currencyCodes.stream().map(c -> {
                 Double rate = rates.get(c.toString());
 
-                return new Currency(c.toString(), rate, currencyCode, timestamp);
+                return new Currency(c.toString(), BigDecimal.valueOf(rate), currencyCode, timestamp);
             }).collect(Collectors.toSet());
 
             currencyRepository.saveAll(currencySet);
